fix insert observation issue
when using composit foreign keys for the instrument
the relationship
instrument = db.relationship('Instrument')
was resulting in all SQL INSERTS to use the value of 'stroubles1' for
the site_id, even when site_id of the inserted object had a different
value. Not sure why this was the case, but the instrument relationship
wasn't really being used anyway.